second commit

+++ level 09/classwork/cw.py
@@ -1,6 +1,6 @@
 #შემოატანინეთ მომხმარებელს ორი input, რომელსაც float() ფუნქციის გამოყენებით გადააქცევთ ათწილადად, საბოლოოდ მოახდინეთ მათემატიკური ოპერაციები ამ ორ რიცხვს შორის
-number = int(input("text a number: "))
-number2 = int(input("text a number: "))
+number = float(input("text a number: "))
+number2 = float(input("text a number: "))
 print(number + number2)
 print(number - number2)
 print(number * number2)
